fix indices on topics.iter
