+++ core/src/execution/proof.rs
@@ -2,7 +2,7 @@ use std::collections::HashMap;
 
 use alloy::primitives::{keccak256, Bytes, B256, U256};
 use alloy::rlp;
-use alloy::rpc::types::{EIP1186AccountProofResponse, Filter, Log};
+use alloy::rpc::types::EIP1186AccountProofResponse;
 use alloy_trie::{
     proof::verify_proof,
     {Nibbles, TrieAccount},
@@ -86,52 +86,4 @@ fn is_empty_value(value: &[u8]) -> bool {
     let is_empty_slot = value.len() == 1 && value[0] == 0x80;
     let is_empty_account = value == empty_account || value == new_empty_account;
     is_empty_slot || is_empty_account
-}
-
-/// Ensure that each log entry in the given array of logs match the given filter.
-pub fn ensure_logs_match_filter(logs: &[Log], filter: &Filter) -> Result<()> {
-    fn log_matches_filter(log: &Log, filter: &Filter) -> bool {
-        if let Some(block_hash) = filter.get_block_hash() {
-            if log.block_hash.unwrap() != block_hash {
-                return false;
-            }
-        }
-        if let Some(from_block) = filter.get_from_block() {
-            if log.block_number.unwrap() < from_block {
-                return false;
-            }
-        }
-        if let Some(to_block) = filter.get_to_block() {
-            if log.block_number.unwrap() > to_block {
-                return false;
-            }
-        }
-        if !filter.address.matches(&log.address()) {
-            return false;
-        }
-        for (i, topic) in filter
-            .topics
-            .iter()
-            .filter(|topic| !topic.is_empty())
-            .enumerate()
-        {
-            if let Some(log_topic) = log.topics().get(i) {
-                if !topic.matches(log_topic) {
-                    return false;
-                }
-            } else {
-                // if filter topic is not present in log, it's a mismatch
-                return false;
-            }
-        }
-        true
-    }
-
-    for log in logs {
-        if !log_matches_filter(log, filter) {
-            return Err(ExecutionError::LogFilterMismatch().into());
-        }
-    }
-
-    Ok(())
 }
